integrate shape into shapeBuilder
